+++ src/main/java/edu/kis/vh/nursery/DefaultCountingOutRhymer.java
@@ -1,6 +1,6 @@
 package edu.kis.vh.nursery;
 
-import edu.kis.vh.nursery.list.IntLinkedList;
+import edu.kis.vh.nursery.stack.IntLinkedList;
 
 public class DefaultCountingOutRhymer implements Stackable{
